Сделала получение комментов к офферу в OfferController с возможностью передать в строке запроса лимит на количество комментариев. То же для списка офферов. Сделала CommentController, добавление комментария.

+++ src/libs/models/offer/offer.controller.ts
@@ -1,9 +1,8 @@
 import { inject, injectable } from 'inversify';
 import { Request, Response } from 'express';
-import { BaseController, HttpMethod } from '../../../rest/index.js';
+import { BaseController, HttpMethod, HttpError, RequestQuery } from '../../../rest/index.js';
 import { Component } from '../../../types/component-enum.js';
 import { Logger } from '../../logger/logger.interface.js';
-import { HttpError } from '../../../rest/index.js';
 import { OfferService } from './offer-service.interface.js';
 import { fillDTO } from '../../../helpers/common.js';
 import { OfferRdo } from './index.js';
@@ -12,12 +11,15 @@ import { StatusCodes } from 'http-status-codes';
 import { CreateOfferRequest } from './types/create-offer-request.type.js';
 import { ParamOfferId } from './types/param-offerid.type.js';
 import { UpdateOfferDto } from './dto/update-offer.dto.js';
+import { CommentService } from '../comment/comment-service.interface.js';
+import { CommentRdo } from '../comment/rdo/comment.rdo.js';
 
 @injectable()
 export class OfferController extends BaseController {
   constructor(
     @inject(Component.Logger) protected readonly logger:Logger,
     @inject(Component.OfferService) private readonly offerService: OfferService,
+    @inject(Component.CommentService) private readonly commentService: CommentService
   ){
     super(logger);
     this.logger.info('Register routes for OfferController...');
@@ -26,10 +28,11 @@ export class OfferController extends BaseController {
     this.addRoute({ path: '/:offerID', method: HttpMethod.GET, handler: this.showDetails });
     this.addRoute({ path: '/:offerID', method: HttpMethod.DELETE, handler: this.delete });
     this.addRoute({ path: '/:offerID', method: HttpMethod.PATCH, handler: this.update });
+    this.addRoute({ path: '/:offerID/comments', method: HttpMethod.GET, handler: this.getComments });
   }
 
-  public async index(_req: Request, res: Response): Promise<void> {
-    const offers = await this.offerService.find();
+  public async index(req: Request, res: Response): Promise<void> {
+    const offers = await this.offerService.find(Number(req.query.limit));
     const responseData = fillDTO(OfferRdo, offers);
     this.ok(res, responseData);
   }
@@ -54,7 +57,15 @@ export class OfferController extends BaseController {
   }
 
   public async delete (req: Request, res: Response): Promise<void> {
+    if (! await this.offerService.exists(req.params.offerID)) {
+      throw new HttpError(
+        StatusCodes.NOT_FOUND,
+        `Offer with id «${req.params.offerID}» not found.`,
+        'OfferController'
+      );
+    }
     const offerDeleted = await this.offerService.deleteById(req.params.offerID);
+    await this.commentService.deleteByOfferID(req.params.offerID);
     if (!offerDeleted) {
       throw new HttpError(
         StatusCodes.NOT_FOUND,
@@ -68,7 +79,7 @@ export class OfferController extends BaseController {
 
   public async update ({ body, params }: Request<ParamOfferId, unknown, UpdateOfferDto>, res: Response): Promise<void> {
     const offerUpdate = await this.offerService.updateById(params.offerID, body);
-    if (!offerUpdate) { // как узнать, почему именно сервер не вернул данные? Надо обработать разные ошибки.
+    if (!offerUpdate) {
       throw new HttpError(
         StatusCodes.NOT_FOUND,
         `Offer with id «${params.offerID}» not found.`,
@@ -76,6 +87,20 @@ export class OfferController extends BaseController {
       );
     }
     const responseData = fillDTO(OfferDetailsRdo, offerUpdate);
+    this.ok(res, responseData);
+  }
+
+  public async getComments ({ params, query }: Request<ParamOfferId, unknown, unknown, RequestQuery>, res: Response): Promise<void> {
+    if (! await this.offerService.exists(params.offerID)) {
+      throw new HttpError(
+        StatusCodes.NOT_FOUND,
+        `Offer with id «${params.offerID}» not found.`,
+        'OfferController'
+      );
+    }
+    console.log('limit value - ', query.limit);
+    const comments = await this.commentService.findByOfferId(params.offerID, Number(query.limit));
+    const responseData = fillDTO(CommentRdo, comments);
     this.ok(res, responseData);
   }
 }
